Add tests for error way with small refactoring
